# server/openapi.yaml
openapi: 3.0.3
info:
  title: AIBillingSystem API
  description: |
    API for the AIBillingSystem - a comprehensive medical billing and practice management system
    designed for therapy practices. Features include patient management, claims processing,
    appointment scheduling, analytics, and AI-powered billing optimization.

    ## Authentication
    Most endpoints require authentication via session-based auth (Replit Auth).
    Protected endpoints will return 401 Unauthorized if not authenticated.

    ## Multi-Tenancy
    The system supports multiple practices. Users are assigned to a practice and can only
    access data within their authorized practice (except admin users who have cross-practice access).

    ## HIPAA Compliance
    This API handles Protected Health Information (PHI). All access is logged for audit purposes.
  version: 1.0.0
  contact:
    name: AIBillingSystem Support
  license:
    name: Proprietary
servers:
  - url: /api
    description: API base path
...
paths:
...
  /auth/user:
    get:
      tags:
        - Authentication
      summary: Get current authenticated user
      description: Returns the currently authenticated user's profile information
      operationId: getCurrentUser
      security:
        - sessionAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /users:
    get:
      tags:
        - Authentication
      summary: List all users (Admin only)
      description: Returns a list of all users in the system. Requires admin role.
      operationId: listUsers
      security:
        - sessionAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{id}/role:
    patch:
      tags:
        - Authentication
      summary: Update user role (Admin only)
      description: Updates a user's role. Cannot remove your own admin role.
      operationId: updateUserRole
      security:
        - sessionAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [therapist, admin, billing]
                  description: New role for the user
      responses:
        '200':
          description: Role updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ==================== PATIENTS ====================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: connect.sid
      description: Session-based authentication via Replit Auth
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Unauthorized
    Forbidden:
      description: Access denied - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Access denied. Admin role required.
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Invalid request parameters
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: Resource not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: An unexpected error occurred
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Error message
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user ID
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        profileImageUrl:
          type: string
          format: uri
        practiceId:
          type: integer
        role:
          type: string
          enum: [therapist, admin, billing]
        credentials:
          type: string
          description: Professional credentials (e.g., OTR/L)
        licenseNumber:
          type: string
        npiNumber:
          type: string
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSummary:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [therapist, admin, billing]
        createdAt:
          type: string
          format: date-time
...
security:
  - sessionAuth: []